requirements.txt
Django==2.2.2
django-crispy-forms==1.7.2
django-daterange-filter==1.3.0
django-easy-pdf==0.1.1
django-multiselectfield==0.1.8
django-mysql==3.0.0.post1
html5lib==1.0.1
lxml==4.3.3
Markdown==3.1
mysql-connector-python==8.0.16
mysqlclient==1.4.2
Pillow==6.0.0
PyPDF2==1.26.0
python-docx==0.8.10
pytz==2019.1
reportlab==3.5.23
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
virtualenv==16.6.0
webencodings==0.5.1
wfastcgi==3.0.0
xhtml2pdf==0.2.3
